% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{appendList}
\alias{appendList}
\title{Append two lists}
\usage{
appendList(x, y)
}
\arguments{
\item{x}{List base}

\item{y}{Second list}
}
\value{
x
}
\description{
Get commom parameters in two list and generate one append list
}
