% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{buildQ}
\alias{buildQ}
\title{Build a dependence matrix}
\usage{
buildQ(Ws, Wt, tau, rho_s = 0.99, rho_t = 0.8, rho_st = 0)
}
\arguments{
\item{tau}{Vector of precisions of variables}

\item{rho_s}{Spatial dependence parameter}

\item{rho_t}{Temporal dependence parameter}

\item{rho_st}{Spatio-temporal dependence parameter}

\item{S}{A neighborhood matrix}
}
\value{
Q a dependence structure
}
\description{
Build a multivariate dependence structure
}
