% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildQC_cpp}
\alias{buildQC_cpp}
\title{buildQC_cpp}
\usage{
buildQC_cpp(M, W, rho)
}
\arguments{
\item{M}{A vector containing the number of neighbors}

\item{W}{A matrix containing the neighborhood structure}

\item{rho}{The spatial dependence parameter}
}
\description{
CAR covariance matrix
}
