% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{buildQST_cpp}
\alias{buildQST_cpp}
\title{buildQST_cpp}
\usage{
buildQST_cpp(Q, Ws, Wt, rho_s, rho_t, rho_st)
}
\arguments{
\item{Ws}{Spatial neighborhod matrix}

\item{Wt}{Temporal neighborhod matrix}

\item{rho_s}{Spatial dependence parameter}

\item{rho_t}{Temporal dependence parameter}

\item{rho_st}{Spatio-temporal dependence parameter}

\item{D}{A vector containing the number of neighbors on diagonal}
}
\description{
Build a Spatio-temporal covariance matrix
}
