% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.tgmrf}
\alias{coef.tgmrf}
\title{Coefficients for tgmrf_model class}
\usage{
\method{coef}{tgmrf}(object)
}
\arguments{
\item{object}{tgmrf_model object to restore the coefficients}
}
\description{
Coefficients for tgmrf_model class
}
