% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CPO_Poisson}
\alias{CPO_Poisson}
\title{Conditional Predictive Ordinate for Poisson data}
\usage{
CPO_Poisson(y, lambda)
}
\arguments{
\item{y}{Vector of response variable}

\item{lambda}{Vector of intensities for each region where y was avaliable}
}
\value{
cpo Conditional Predictive Ordinate
}
\description{
Conditional Predictive Ordinate for Poisson data
}
