% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_beta_conj_pois_cpp}
\alias{dens_beta_conj_pois_cpp}
\title{dens_beta_conj_pois_cpp}
\usage{
dens_beta_conj_pois_cpp(x, params, eps, nu, DSigma)
}
\arguments{
\item{x}{Point to restore the density}

\item{params}{}
}
\description{
To usage in MCMC estimation
}
