% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dens_eps_pois_cpp}
\alias{dens_eps_pois_cpp}
\title{dens_eps_pois_cpp}
\usage{
dens_eps_pois_cpp(x, params, ele, eps, nu, Q, DSigma, Xbeta)
}
\arguments{
\item{x}{Point to restore the density}

\item{params}{A list with some informations about the MCMC model.}
}
\description{
To usage in MCMC estimation
}
