% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(y, lambda)
}
\arguments{
\item{y}{Vector of response variable}

\item{lambda}{Matrix of intensities for each region where y was avaliable in each MCMC iteration}
}
\value{
DIC Deviance Information Criterion measure
}
\description{
Deviance Information Criterion
}
