% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dtruncnorm}
\alias{dtruncnorm}
\title{dtruncnorm}
\usage{
dtruncnorm(a, b, mean, var, x)
}
\arguments{
\item{a}{Lower bound}

\item{b}{Upper bound}

\item{mean}{Normal expectation}

\item{var}{Normal variance}

\item{x}{Point to evaluate the density}
}
\description{
To restore a point density of a truncated normal
}
