% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{erg_mean}
\alias{erg_mean}
\title{Ergodic mean}
\usage{
erg_mean(x)
}
\arguments{
\item{x}{Vector to restore the ergodic mean}
}
\value{
erg A vector with the ergodic mean
}
\description{
Return the ergodic mean of a vector
}
