% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LPML}
\alias{LPML}
\title{Log Pseudo Marginal Likelihood}
\usage{
LPML(y, lambda)
}
\arguments{
\item{y}{Vector of response variable}

\item{lambda}{Matrix of intensities for each region where y was avaliable in each MCMC iteration}
}
\value{
LPML -2 * Log Pseudo Marginal Likelihood measure
}
\description{
Log Pseudo Marginal Likelihood
}
