% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_range}
\alias{max_range}
\title{Maximum range dependence parameters}
\usage{
max_range(Ws, Wt, rho_s = NULL, rho_t = NULL, rho_st = NULL)
}
\arguments{
\item{rho_s}{Spatial dependence parameter}

\item{rho_t}{Temporal dependence parameter}

\item{rho_st}{Spatio-temporal dependence parameter}

\item{S}{A neighborhood matrix}

\item{tau}{Vector of precisions of variables}
}
\value{
Q a dependence structure
}
\description{
Find the maximum range of the dependence parameters (comparing with the parameters passed)
}
