% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_range_cpp}
\alias{max_range_cpp}
\title{max_range_cpp}
\usage{
max_range_cpp(Ws, Wt, rho_s, rho_t, rho_st)
}
\arguments{
\item{Ws}{Spatial adjacency matrix}

\item{Wt}{Temporal adjacency matrix}

\item{rho_s}{Spatial parameter}

\item{rho_t}{Temporal parameter}

\item{rho_st}{Spatio-temporal parameter}
}
\description{
Max range to sample rho using rcpp
}
