% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_poisson_st.R
\name{mcmc_poisson_st}
\alias{mcmc_poisson_st}
\title{MCMC poisson for spatio-temporal data}
\usage{
mcmc_poisson_st(y, X, E, n_reg, n_var, neigh, beta, nu, eps, rho_s, rho_t,
  rho_st, tau, nsim, burnin, thin, type, type_num, mat_type, method,
  tau_beta, mean_beta, eta_nu, psi_nu, ninit, maxpoint, var_beta, var_eps,
  var_log_nu, var_rho, fix_rho_s, fix_rho_t, fix_rho_st, range_rho_s,
  range_rho_t, range_rho_st, verbose, c_beta, c_eps, c_nu, c_rho)
}
\arguments{
\item{y}{Response variable}

\item{X}{Matrix or data.frame of covariates}

\item{E}{Offsets for Poisson data}

\item{n_reg}{Number of regions in study}

\item{n_var}{Number of variables in study}

\item{neigh}{Neighborhood structure of nb class or a neighborhood matrix}

\item{beta}{Vector of the initial values to coeficients vector}

\item{nu}{A Initial value to variance of copula}

\item{eps}{Initial values to eps parameters}

\item{rho_s}{A Initial value to spatial dependence}

\item{rho_t}{A Initial value to temporal dependence}

\item{rho_st}{A Initial value to spatio-temporal dependence}

\item{tau}{Precision parameter vector of MCAR}

\item{nsim}{Number of MCMC iterations}

\item{burnin}{number of discards iterations}

\item{thin}{Lag to collect the observations}

\item{type}{'log-normal', 'gamma-shape', 'gamma-scale', 'weibull-shape', 'weibull-scale'}

\item{type_num}{A numeric that represent type in .C functions}

\item{mat_type}{car or leroux}

\item{method}{arms or metropolis}

\item{tau_beta}{Prior precision of beta vector}

\item{mean_beta}{Prior mean of beta vector}

\item{eta_nu}{Prior Shape of nu parameter}

\item{psi_nu}{Prior Scale of nu parameter}

\item{ninit}{Number of initial points in ARMS method}

\item{maxpoint}{Maximum number of evaluation a envelope in each iteration}

\item{var_beta}{Variance of beta proposal (metropolis)}

\item{var_eps}{Variance of beta proposal (metropolis)}

\item{var_rho}{Variance matrix (metropolis)}

\item{fix_rho_s}{Is rho_s fixed?}

\item{fix_rho_t}{Is rho_t fixed?}

\item{fix_rho_st}{Is rho_st fixed?}

\item{range_rho_s}{Range to sample rho_s}

\item{range_rho_t}{Range to sample rho_t}

\item{range_rho_st}{Range to sample rho_st}
}
\value{
eps A matrix with eps samples

mu A matrix with intensities samples

beta A matrix with beta samples

nu A vector with samples of nu

rho A vector with samples of rho

model_info A data_frame with some models' informations

fit measures A data_frame with some fit measures
}
\description{
MCMC for Poisson data using TGMRF
}
