% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MEASURES}
\alias{MEASURES}
\title{LPML, DIC and WAIC measures}
\usage{
MEASURES(y, lambda)
}
\arguments{
\item{y}{Vector of response variable}

\item{lambda}{Matrix of intensities for each region where y was avaliable in each MCMC iteration}
}
\value{
measures A data.frame with LMPL, DIC and WAIC measures
}
\description{
LPML, DIC and WAIC measures
}
