% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mode}
\alias{mode}
\title{Mode based on the empirical density}
\usage{
mode(x, bw = 0.1)
}
\arguments{
\item{x}{Numerical vector}

\item{bw}{The smoothing bandwidth to be used}
}
\value{
m Mode
}
\description{
Mode based on the empirical density
}
