% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nenia_s-data.R
\docType{data}
\name{nenia_s}
\alias{nenia_s}
\title{Nenia Tridens counts}
\format{An object of class \code{"data.frame"}}
\usage{
data(nenia_s)
}
\description{
Data from a spatial experiment with gastropods of nenia tridens specie.
The data was collect in a forest located in Costa Rica.#'
}
\keyword{datasets}
