% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nenia_st-data.R
\docType{data}
\name{nenia_st}
\alias{nenia_st}
\title{Nenia Tridens counts for spatio-temporal analisys}
\format{An object of class \code{"data.frame"}}
\usage{
data(nenia_s)
}
\description{
Data from a spatio-temporal experiment with gastropods of nenia tridens specie.
The data was collect in a forest located in Costa Rica.
}
\keyword{datasets}
