% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.tgmrf}
\alias{plot.tgmrf}
\title{Plot for tgmrf_model class}
\usage{
\method{plot}{tgmrf}(object, series = T, series3D = F, erg = F,
  boxplot = F, coord = NULL, xlim = NULL, ylim = NULL,
  col = "#4682b460", pch = 20, main = "Spatial effects", bty = "l",
  ask = T, ...)
}
\arguments{
\item{object}{tgmrf_model object to plot}
}
\description{
Plot for tgmrf_model class
}
