% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Plot MCMC}
\usage{
plotMCMC(chain, xlab = "x", ylab = "Iteration", zlab = "Density",
  alpha = 0.5, col = grey(0.2), bty = "b2", phi = 20, theta = 50,
  ticktype = "detailed", lwd_density = 3, lwd_chain = 1, ...)
}
\arguments{
\item{chain}{Vector to restore the ergodic mean}
}
\value{
erg A vector with the ergodic mean
}
\description{
Plot MCMC chains in 3D
}
