% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tgmrf}
\alias{print.tgmrf}
\title{Print for tgmrf_model class}
\usage{
\method{print}{tgmrf}(object)
}
\arguments{
\item{object}{tgmrf_model object to print}
}
\description{
Print for tgmrf_model class
}
