% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PSEUDO_R2}
\alias{PSEUDO_R2}
\title{Pseudo R square}
\usage{
PSEUDO_R2(y, lambda, residual = "raw")
}
\arguments{
\item{y}{Vector of response variable}

\item{lambda}{Matrix of intensities for each region where y was avaliable in each MCMC iteration}
}
\value{
R2 Pseudo R2 based on raw residuals or deviance residuals
}
\description{
Pseudo R square
}
