% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm_cpp}
\alias{rmvnorm_cpp}
\title{rmvnorm_cpp}
\usage{
rmvnorm_cpp(sigma)
}
\arguments{
\item{sigma}{covariance matrix}

\item{n}{number of samples}

\item{mean}{vector of means}
}
\description{
Random values from a multivariate normal
}
