% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgmrf_st.R
\name{rtgmrf_st}
\alias{rtgmrf_st}
\title{Random observations of a Transformed Gaussian Markov Random Field}
\usage{
rtgmrf_st(rowid = 10, colid = 10, n_var = 2, X = NULL,
  neigh = NULL, n_trials = NULL, E = NULL, rho_s = 0.9,
  rho_t = 0, rho_st = 0, betas = c(-0.1, 0.3, 0.8), intercept = T,
  nu = 2, tau = 1, type_data = "gamma-shape", family = "poisson",
  mat_type = "car", seed = 1)
}
\arguments{
\item{rowid}{Number of lines of a lattice. (Used if neigh = NULL)}

\item{colid}{Number of columns of a lattice. (Used if neigh = NULL)}

\item{n_var}{Number of variables in problem (for multivariate problems)}

\item{neigh}{A object of class nb}

\item{rho_s}{Spatial dependence parameter}

\item{rho_t}{Temporal dependence parameter}

\item{rho_st}{Spatio-temporal dependence parameter}

\item{betas}{Coeficients}

\item{nu}{Dispersion parameter for gamma models}

\item{tau}{Vector of precisions of variables}

\item{type_data}{Depends of family.
'lognormal', 'lognormal-precision', 'gamma-shape', 'gamma-scale', 'weibull-shape', 'weibull-scale' for Poisson family
'beta-logit', 'beta-probit', 'beta-alpha', 'beta-beta' for Binomial family}

\item{family}{'poisson' or 'binary'}

\item{seed}{A seed to reproduce the reuslts}

\item{n}{Vector with size of binomial trials}
}
\value{
y Response variable

X Covariates matrix

neigh Neighborhood structure

Q Covariance matrix

mu Means (TGMRF)

eps Errors (GMRF)
}
\description{
Use this function to simulate a data of a spatio-temporal TGMRF.
}
