% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_tgmrf.R
\name{st_tgmrf}
\alias{st_tgmrf}
\title{Spatio-temporal Transformed Gaussian Markov Random Field}
\usage{
st_tgmrf(y, X, n_reg, n_var, beta, nu, eps, mu, rho_s, rho_t, rho_st, tau,
  family, type, mat_type, method, nsim, burnin, thin, E, n, neigh,
  prior_param, MCMC_config, fix_rho, range, verbose, c_beta, c_eps, c_mu,
  c_nu, c_rho)
}
\arguments{
\item{y}{Response variable}

\item{X}{Matrix or data.frame of covariates}

\item{n_reg}{Number of regions in study}

\item{n_var}{Number of variables in study}

\item{beta}{Vector of the initial values to coeficients vector}

\item{nu}{A Initial value to variance of copula}

\item{rho_s}{Spatial dependence parameter}

\item{rho_t}{Temporal dependence parameter}

\item{rho_st}{Spatio-temporal dependence parameter}

\item{tau}{Vector of precisions of variables}

\item{family}{'Poisson' or 'Binary'}

\item{type}{Depends of family.
'lognormal', 'lognormal-precision', 'gamma-shape', 'gamma-scale', 'weibull-shape', 'weibull-scale' for Poisson family
'beta-logit', 'beta-probit', 'beta-alpha', 'beta-beta' for Binomial family}

\item{mat_type}{car or leroux}

\item{method}{'arms' or 'metropolis'}

\item{nsim}{Number of MCMC iterations}

\item{burnin}{number of discards iterations}

\item{thin}{Lag to collect the observations}

\item{E}{Offsets for Poisson data}

\item{n}{Number of trials in Binomial data}

\item{neigh}{Neighborhood structure of nb class or a neighborhood matrix}

\item{prior_param}{List of priors parameters. Gaussian for beta vector and IGamma for nu. Entrys are: 'nu' and 'beta'}

\item{MCMC_config}{Parameters to tunning MCMC}

\item{fix_rho}{A list informing if any rho (rho_s, rho_t, rho_st) is fixed}

\item{range}{Range to use during the MCMC. Entrys are: 'rho_s', 'rho_t' and 'rho_st'}

\item{range}{A list informing a range for sampling rho (rho_s, rho_t, rho_st). Each entre must be a vector in the interval (-1, 1).}
}
\value{
beta A matrix with samples of beta

nu A vector with samples of nu

rho A vector with samples of rho
}
\description{
Uses this function to fit spatio-temporal data using non-guassian copulas.
}
\examples{
s_tgmrf(1,1,1,1)

}
