% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.tgmrf}
\alias{summary.tgmrf}
\title{Summary for tgmrf_model class (coda is necessary)}
\usage{
\method{summary}{tgmrf}(object, bw = 0.1, HPD = T, level = 0.95)
}
\arguments{
\item{object}{tgmrf_model object to summarise}
}
\description{
Summary for tgmrf_model class (coda is necessary)
}
