% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgmrf-class.R
\docType{class}
\name{tgmrf-class}
\alias{tgmrf-class}
\alias{tgmrf_class}
\title{An object of 'tgmrf' class}
\description{
An object of 'tgmrf' class
}
\section{Slots}{

\describe{
\item{\code{eps}}{A matrix with errors samples in each region}

\item{\code{mu}}{A matrix with means samples in each region}

\item{\code{beta}}{A matrix with coefficients samples}

\item{\code{nu}}{A vector with dispersion parameter sample}

\item{\code{rho}}{A vector with spatial parameter sample}

\item{\code{model_info}}{A list with other informations about the model}

\item{\code{fit_measures}}{A data.frame with fit measures}

\item{\code{time_elapsed}}{A numeric with time elapsed running}

\item{\code{call}}{A call that result in this object}
}}

